Create a basic game architecture.
This commit defines a basic environment in which this game will be developed.
This environment is based on the idea of separating game data from the game
loop.  Therefore, the classes that update every frame are kept completely
distinct from the classes that make up the game state.

The user interface loop is by far the most developed in this commit.  Expect
later versions to include player loops and a central messaging system.