Add a minimal messaging system.
Eventually, I want to build the messaging system up to the point where it can
make the network transparent to the various game loops that are running.
However, this commit only creates the most minimal system possible.  Messages
can only be sent between loops on the same machine.

I am still not sure if the messenger should be based on callbacks or message
stacks.  Callbacks seems to make the messenger less dependent on the rest of
the architecture, while stacks could be more transparent and thread-safe.