Expand the user interface code.
This commit includes improvements to a number of components of the GUI.  First
of all, I introduced more concrete roles for the style & geometry helpers and
the artists.  The geometry helper is simply responsible to generating shapes.
The style helper is still responsible for providing color information, but it
is now also responsible for providing information that could influence what
shape the geometry helper creates.  The artists are responsible for combining
this information and actually rendering the game objects.

I also introduced a layers helper that manages the order in which artists draw
themselves.  The draw method for each artist is now called several times each
frame, and the artists are responsible for making sure that they only draw
when the right layer is being edited.

Finally, I added a few methods to the geometry helper that produce new, useful
shapes.  These shapes are circles, hexagons, and hollow hexagons.